added mail sending
